3DMM-Fitting-Pytorch/requirements.txt
autocommand==2.2.2
backports.tarfile==1.2.0
brotli==1.0.9
certifi==2020.6.20
charset-normalizer==3.3.2
cycler==0.10.0
decorator==4.4.2
face-alignment==1.1.1
facenet-pytorch==2.5.2
filelock==3.17.0
fsspec==2025.3.2
future==0.18.2
fvcore==0.1.2.post20201111
gmpy2==2.2.1
idna==3.7
imagecodecs==2024.9.22
imageio-ffmpeg==0.6.0
imageio==2.9.0
importlib-metadata==8.6.1
inflect==7.3.1
jaraco.collections==5.1.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
jinja2==3.1.6
kiwisolver==1.3.1
lazy-loader==0.4
markupsafe==3.0.2
matplotlib==3.3.3
mkl-fft==1.3.11
mkl-random==1.2.8
mkl-service==2.4.0
more-itertools==10.3.0
mpmath==1.3.0
networkx==2.5
numpy==1.19.4
opencv-python==4.4.0.46
packaging==25.0
pillow==8.0.1
pip==25.0
platformdirs==4.2.2
portalocker==2.0.0
pyparsing==2.4.7
pysocks==1.7.1
python-dateutil==2.8.1
pytorch3d==0.7.8
pywavelets==1.1.1
pyyaml==5.3.1
requests==2.32.3
scikit-image==0.19.3
scipy==1.5.4
setuptools==75.8.0
six==1.15.0
sympy==1.13.3
tabulate==0.8.7
termcolor==1.1.0
tifffile==2020.11.18
tomli==2.0.1
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tqdm==4.52.0
triton==3.0.0
typeguard==4.3.0
typing-extensions==4.12.2
urllib3==2.3.0
wheel==0.45.1
yacs==0.1.8
zipp==3.21.0